+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Catch errors during pings to Google Analytics, for example in case of a 
   missing internet connection
+- Track additional browser data such as screen size and language
 
 ## 0.0.4